requirements.txt
requests
beautifulsoup4
openai
langchain-openai
python-dotenv